feature cliente e endereco 100% prontos.

+++ E-commerce/Application/ClienteService/ClienteService.cs
@@ -1,15 +1,11 @@
 ﻿using Domain.Interfaces;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Application.Interfaces;
 using Application.Dtos;
 using AutoMapper;
 using Domain.Entities;
 using Application.ClienteExceptions;
-using System.Linq.Expressions;
 
 namespace Application.ClienteService
 {
@@ -25,6 +21,7 @@ namespace Application.ClienteService
             _validacoes = validacoes;
             _mapper = mapper;
         }
+
         public void CadastrarCliente(CadastroClienteDto dto)
         {
             _validacoes.ValidarNome(dto);
@@ -32,53 +29,63 @@ namespace Application.ClienteService
             _validacoes.ValidarSenha(dto);
             _validacoes.ValidarCpf(dto);
 
+            Cliente clienteCpf = _repository.BuscarCpf(dto.Cpf);
+            Cliente clienteEmail = _repository.BuscarEmail(dto.Email);
+
+            if (clienteCpf != null || clienteEmail != null)
+                throw new ClienteExistente();
 
             Cliente cliente = _mapper.Map<Cliente>(dto);
-
-            BuscarCpfEmail(dto.Email, dto.Cpf);
-
             _repository.CadastrarCliente(cliente);
-
         }
+
         public List<ListarClientesDTO> ListarClientes()
         {
             List<Cliente> clientes = _repository.ListarClientes();
-            return _mapper.Map<List<ListarClientesDTO>>(clientes);
+            List<ListarClientesDTO> dtoList = _mapper.Map<List<ListarClientesDTO>>(clientes);
+            return dtoList;
         }
+
         public void RemoverCliente(RemoverClienteDTO dto)
         {
-            Cliente clienteBuscado = BuscarEmail(dto.Email);
+            Cliente clienteBuscado = _repository.BuscarEmail(dto.Email);
+
+            if (clienteBuscado == null)
+                throw new ClienteNaoExiste();
 
             if (clienteBuscado.Senha != dto.Senha)
-            {
                 throw new SenhaIncorreta();
-            }
 
             _repository.RemoverCliente(clienteBuscado);
         }
+
         public void AlterarNome(NovoNomeClienteDTO dto)
         {
             _validacoes.ValidarNovoNome(dto);
 
-            Cliente clienteBuscado = BuscarCpf(dto.Cpf);
+            Cliente clienteBuscado = _repository.BuscarCpf(dto.Cpf);
+
+            if (clienteBuscado == null)
+                throw new ClienteNaoExiste();
 
             if (clienteBuscado.Senha != dto.Senha)
-            {
                 throw new SenhaIncorreta();
-            }
 
             _repository.AlterarNome(clienteBuscado, dto.NovoNome);
         }
+
         public void AlterarSenha(AlterarSenhaDTO dto)
         {
             _validacoes.ValidarNovaSenha(dto);
 
-            Cliente clienteBuscado = BuscarEmail(dto.Email);
+            Cliente clienteBuscado = _repository.BuscarEmail(dto.Email);
+
+            if (clienteBuscado == null)
+                throw new ClienteNaoExiste();
 
             if (clienteBuscado.Senha != dto.Senha)
-            {
                 throw new SenhaIncorreta();
-            }
+
             _repository.AlterarSenha(clienteBuscado, dto.NovaSenha);
         }
 
@@ -86,69 +93,30 @@ namespace Application.ClienteService
         {
             _validacoes.ValidarNovoEmail(dto);
 
-            Cliente clienteBuscado = BuscarEmail(dto.Email);
+            Cliente clienteBuscado = _repository.BuscarEmail(dto.Email);
+
+            if (clienteBuscado == null)
+                throw new ClienteNaoExiste();
 
             if (clienteBuscado.Senha != dto.Senha)
-            {
                 throw new SenhaIncorreta();
-            }
+
+            Cliente emailNovoBuscado = _repository.BuscarEmail(dto.NovoEmail);
+            if (emailNovoBuscado != null)
+                throw new ClienteExistente();
 
             _repository.AlterarEmail(clienteBuscado, dto.NovoEmail);
         }
 
         public BuscarClienteSaidaDTO BuscarClienteEspecifico(BuscarClienteEntradaDTO dto)
         {
-            Cliente ClienteBuscado = BuscarCpf(dto.Cpf);
+            Cliente clienteBuscado = _repository.BuscarCpf(dto.Cpf);
 
-           BuscarClienteSaidaDTO clientesaidaDTO = _mapper.Map<BuscarClienteSaidaDTO>(ClienteBuscado);
-
-            return clientesaidaDTO;
-        }
-
-        public Cliente BuscarCpf(string cpf)
-        {
-            Cliente clientecpf = _repository.BuscarCpf(cpf);
-
-            if (clientecpf == null)
-            {
+            if (clienteBuscado == null)
                 throw new ClienteNaoExiste();
-            }
-            else
-            {
-                return clientecpf;
-            }
 
-
+            BuscarClienteSaidaDTO dtoSaida = _mapper.Map<BuscarClienteSaidaDTO>(clienteBuscado);
+            return dtoSaida;
         }
-
-        public Cliente BuscarEmail(string email)
-        {
-            Cliente clienteemail = _repository.BuscarEmail(email);
-
-            if (clienteemail == null)
-            {
-                throw new ClienteNaoExiste();
-            }
-            else
-            {
-                return clienteemail;
-            }
-        }
-        public void BuscarCpfEmail(string email, string cpf)
-        {
-            Cliente clientecpf = _repository.BuscarCpf(cpf);
-            Cliente clienteemail = _repository.BuscarEmail(email);
-
-            if (clientecpf != null || clienteemail != null)
-            {
-                throw new ClienteExistente();
-            }
-
-
-        }
-
     }
-
 }
-
-
